Base files - no assets
The EXE example build is EXPLORE.C
The main library code is grid2.h

There's not a lot of structure in the files or anything.  It's just designed to do what it's supposed to do.

Yet this is a project in it's very early stages and should be looked as that.
Not much should be expected out of this.

Oh, and it uses a few RayLib facilities in order to work.

My recommendation would be to use this code to study it.  Maybe comment on it, give some advices / recommendations.

I'm really open for ideas.